Refactor : File Convention states/client
