--- src/recoil/atoms/activeCategory.ts
@@ -1,6 +1,0 @@
-import { atom } from "recoil";
-
-export const activeCategory = atom<string>({
-  key: "activeCategory",
-  default: "주식"
-})
