--- src/recoil/atoms/topicUsertype.ts
@@ -1,6 +1,0 @@
-import { atom } from "recoil";
-
-export const topicUserType = atom<string>({
-  key: "topicUserType",
-  default: "선생님"
-})
